[pre-req] Convert Workpad Config and state to TS (#68615)
