+++ x-pack/plugins/canvas/public/components/workpad_header/view_menu/index.ts
@@ -14,13 +14,11 @@ import { State, CanvasWorkpadBoundingBox } from '../../../../types';
 import { fetchAllRenderables } from '../../../state/actions/elements';
 // @ts-ignore Untyped local
 import { setZoomScale, setFullscreen, selectToplevelNodes } from '../../../state/actions/transient';
-// @ts-ignore Untyped local
 import {
   setWriteable,
   setRefreshInterval,
   enableAutoplay,
   setAutoplayInterval,
-  // @ts-ignore Untyped local
 } from '../../../state/actions/workpad';
 import { getZoomScale, canUserWrite } from '../../../state/selectors/app';
 import {
@@ -75,7 +73,7 @@ const mapDispatchToProps = (dispatch: Dispatch) => ({
   },
   doRefresh: () => dispatch(fetchAllRenderables()),
   setRefreshInterval: (interval: number) => dispatch(setRefreshInterval(interval)),
-  enableAutoplay: (autoplay: number) => dispatch(enableAutoplay(autoplay)),
+  enableAutoplay: (autoplay: number) => dispatch(enableAutoplay(!!autoplay)),
   setAutoplayInterval: (interval: number) => dispatch(setAutoplayInterval(interval)),
 });
