+++ x-pack/plugins/canvas/public/state/actions/embeddable.ts
@@ -6,8 +6,7 @@
 
 import { Dispatch } from 'redux';
 import { createAction } from 'redux-actions';
-// @ts-ignore Untyped
-import { createThunk } from 'redux-thunks';
+import { createThunk } from '../../lib/create_thunk';
 // @ts-ignore Untyped Local
 import { fetchRenderable } from './elements';
 import { State } from '../../../types';
